[REFACTOR] record calculet & oftenused calculet
