+++ client/src/components/calculet-block/CalculetBlock.js
@@ -12,9 +12,10 @@ const Wrapper = styled.div`
  *
  * @param {string} srcCode - 계산기 소스코드
  * @param {string} manual - 계산기 설명 마크다운
+ * @param {string} calculetId - 계산기 번호
  * @returns
  */
-function CalculetBlock({ srcCode, manual, calculetId, userEmail }) {
+function CalculetBlock({ srcCode, manual, calculetId }) {
   /**
    * 받아온 html을 넣는 iframe의 크기를 원본 html 크기 맞게 조절하는 함수
    * @param {event object} e
@@ -34,11 +35,7 @@ function CalculetBlock({ srcCode, manual, calculetId, userEmail }) {
         scrolling="no"
         title="calculet"
       />
-      <CalculetManual
-        content={manual}
-        calculetId={calculetId}
-        userEmail={userEmail}
-      />
+      <CalculetManual content={manual} calculetId={calculetId} />
     </Wrapper>
   );
 }
